Some fixes.

+++ cosmos/models/mobility/src/main/java/es/tid/cosmos/mobility/mapreduce/RepbtsSpreadNodebtsMapper.java
@@ -4,6 +4,7 @@ import java.io.IOException;
 
 import com.twitter.elephantbird.mapreduce.io.ProtobufWritable;
 import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.mapreduce.Mapper;
 
 import es.tid.cosmos.mobility.data.MobProtocol.BtsCounter;
@@ -15,11 +16,11 @@ import es.tid.cosmos.mobility.data.NodeBtsDayUtil;
  *
  * @author dmicol
  */
-public class RepbtsSpreadNodebtsMapper extends Mapper<IntWritable,
+public class RepbtsSpreadNodebtsMapper extends Mapper<LongWritable,
         ProtobufWritable<NodeMxCounter>, ProtobufWritable<NodeBtsDay>,
         IntWritable> {
     @Override
-    public void map(IntWritable key, ProtobufWritable<NodeMxCounter> value,
+    public void map(LongWritable key, ProtobufWritable<NodeMxCounter> value,
             Context context) throws IOException, InterruptedException {
         final NodeMxCounter counter = value.get();
         for (BtsCounter bts : counter.getBtsList()) {
